y Y
git-svn-id: svn+ssh://svn.aslforge.ethz.ch/svnroot/avalon@46 05686622-51ee-4c29-b765-5a36481f58a2

+++ windsensor.h
@@ -18,10 +18,10 @@
 
 DDX_STORE_TYPE(WindData,
   struct {
-    float direction;		// In Degree (from -180 to +180 degrees)
-    float speed;		// Windspeed in Knots
-    float temperature;		// Temperature at the Surface
-    float voltage;		// Voltage of the heating
+    double direction;		// In Degree (from -180 to +180 degrees)
+    double speed;		// Windspeed in Knots
+    double temperature;		// Temperature at the Surface
+    double voltage;		// Voltage of the heating
     int uptodate;		// Are Voltage and Temp up to date? (0 = no, not 0 = yes)
   }
 );
